Refactor database seeding and group insertion scripts
