+++ package.json
@@ -16,9 +16,9 @@
     "db:drop": "drizzle-kit drop",
     "db:check": "drizzle-kit check:pg",
     "db:studio": "drizzle-kit studio",
-    "db:seed": "bun ./scripts/db/seed.ts",
-    "db:seed:cleanup": "bun ./scripts/db/seed.ts --cleanup",
-    "db:insert:groups": "bun ./scripts/db/insertCustomGroups.ts",
+    "db:seed": "pnpx esbuild ./scripts/db/seed.ts --bundle --platform=node --outfile=dist/seed.js && node --env-file=.env dist/seed.js",
+    "db:seed:cleanup": "pnpm run db:seed --cleanup",
+    "db:insert:groups": "pnpx esbuild ./scripts/db/insertCustomGroups.ts --bundle --platform=node --format=cjs | node --env-file=.env",
     "test": "jest --runInBand",
     "format": "prettier --check \"src/**/*.{ts,md}\"",
     "format:fix": "prettier --write \"src/**/*.{ts,md}\""
@@ -39,7 +39,6 @@
     "@types/node": "^20.10.3",
     "@typescript-eslint/eslint-plugin": "^6.19.0",
     "@typescript-eslint/parser": "^6.19.0",
-    "bun": "^1.0.22",
     "csv-parser": "^3.0.0",
     "drizzle-kit": "^0.20.13",
     "esbuild": "^0.19.8",
