+++ scripts/db/seed.ts
@@ -3,16 +3,25 @@ import { cleanup, seed } from '../../src/utils/db/seed';
 
 const DEFAULT_DB_CONNECTION_URL = 'postgresql://postgres:secretpassword@localhost:5432';
 
-if (process.argv.includes('--cleanup')) {
-  const dbConnectionUrl = process.env.DB_CONNECTION_URL ?? DEFAULT_DB_CONNECTION_URL;
-  const { dbPool, connection } = createDbPool(dbConnectionUrl, { max: 1 });
-  await cleanup(dbPool);
-  await connection.end();
-  console.log('Cleaned up database');
-} else {
-  const dbConnectionUrl = process.env.DB_CONNECTION_URL ?? DEFAULT_DB_CONNECTION_URL;
-  const { dbPool, connection } = createDbPool(dbConnectionUrl, { max: 1 });
-  await seed(dbPool);
-  await connection.end();
-  console.log('Seeded database');
+async function main() {
+  if (process.argv.includes('--cleanup')) {
+    const dbConnectionUrl = process.env.DB_CONNECTION_URL ?? DEFAULT_DB_CONNECTION_URL;
+    const { dbPool, connection } = createDbPool(dbConnectionUrl, { max: 1 });
+    await cleanup(dbPool);
+    await connection.end();
+    console.log('Cleaned up database');
+  } else {
+    const dbConnectionUrl = process.env.DB_CONNECTION_URL ?? DEFAULT_DB_CONNECTION_URL;
+    const { dbPool, connection } = createDbPool(dbConnectionUrl, { max: 1 });
+    await seed(dbPool);
+    await connection.end();
+    console.log('Seeded database');
+  }
 }
+
+main()
+  .then(() => process.exit(0))
+  .catch((error) => {
+    console.error('Error seeding database:', error);
+    process.exit(1);
+  });
