+++ src/utils/db/seed.ts
@@ -29,6 +29,7 @@ async function seed(dbPool: PostgresJsDatabase<typeof db>) {
 }
 
 async function cleanup(dbPool: PostgresJsDatabase<typeof db>) {
+  await dbPool.delete(db.userAttributes);
   await dbPool.delete(db.votes);
   await dbPool.delete(db.federatedCredentials);
   await dbPool.delete(db.registrationData);
